feat: clean up library structure
wasm contracts don't need wasmparser in their dep graph i don't think
