+++ Cargo.toml
@@ -244,6 +244,7 @@ members = [
   "lib/voyager-plugin",
   "lib/voyager-sdk",
   "lib/voyager-plugin-protocol",
+  "lib/wasm-client-type",
 ]
 
 [workspace.package]
@@ -396,6 +397,7 @@ state-lens-ics23-mpt-light-client = { path = "cosmwasm/ibc-union/lightclient/sta
 
 solidity-slot = { path = "lib/solidity-slot", default-features = false }
 ucs04         = { path = "lib/ucs04", default-features = false }
+wasm-client-type = { path = "lib/wasm-client-type", default-features = false }
 
 # =====================
 # external dependencies
