+++ cosmwasm/ibc-union/core/Cargo.toml
@@ -38,4 +38,4 @@ thiserror       = { workspace = true }
 unionlabs       = { workspace = true, features = ["ethabi", "bincode"] }
 
 [dev-dependencies]
-hex-literal.workspace = true
+hex-literal = { workspace = true }
