+++ lib/wasm-client-type/src/lib.rs
@@ -1,0 +1,78 @@
+use std::{fmt, str::FromStr};
+
+#[doc(hidden)]
+pub use paste::paste;
+use serde::{Deserialize, Serialize};
+
+#[macro_export]
+macro_rules! export_wasm_client_type {
+    ($type:ident) => {
+        const _: $crate::WasmClientType = $crate::WasmClientType::$type;
+        $crate::paste! {
+            #[no_mangle]
+            #[used]
+            #[allow(non_upper_case_globals)]
+            static [ <WASM_CLIENT_TYPE_ $type> ]: u8 = 0;
+        }
+    };
+}
+
+/// This type is used to discriminate 08-wasm light clients.
+///
+/// We need to be able to determine the light client from the light client code itself (not instantiated yet).
+/// Light clients supported by voyager must export a `#[no_mangle] static WASM_CLIENT_TYPE_<TYPE>: u8 = 0` variable.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum WasmClientType {
+    Cometbls,
+    Tendermint,
+}
+
+impl FromStr for WasmClientType {
+    type Err = WasmClientTypeParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "Cometbls" => Ok(WasmClientType::Cometbls),
+            "Tendermint" => Ok(WasmClientType::Tendermint),
+            _ => Err(WasmClientTypeParseError::UnknownType(s.to_string())),
+        }
+    }
+}
+
+impl fmt::Display for WasmClientType {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Cometbls => write!(f, "Cometbls"),
+            Self::Tendermint => write!(f, "Tendermint"),
+        }
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, thiserror::Error)]
+pub enum WasmClientTypeParseError {
+    #[error("unknown wasm client type `{0}`")]
+    UnknownType(String),
+}
+
+pub fn extract_from_wasm(
+    bz: impl AsRef<[u8]>,
+) -> Result<Option<WasmClientType>, WasmClientTypeParseError> {
+    wasmparser::Parser::new(0)
+        .parse_all(bz.as_ref())
+        .find_map(|payload| {
+            payload.ok().and_then(|payload| match payload {
+                wasmparser::Payload::ExportSection(e) => Some(e),
+                _ => None,
+            })
+        })
+        .and_then(|exports| {
+            exports.into_iter().find_map(|export| {
+                export
+                    .ok()
+                    .and_then(|export| export.name.strip_prefix("WASM_CLIENT_TYPE_"))
+            })
+        })
+        .map(str::parse)
+        .transpose()
+}
