+++ voyager/modules/state/cosmos-sdk/Cargo.toml
@@ -14,7 +14,6 @@ workspace = true
 [dependencies]
 clap             = { workspace = true, features = ["derive"] }
 cometbft-rpc     = { workspace = true }
-dashmap          = { workspace = true }
 embed-commit     = { workspace = true }
 ibc-classic-spec = { workspace = true }
 jsonrpsee        = { workspace = true, features = ["macros", "server", "tracing"] }
